+++ plugins/repository-gcs/src/test/java/org/elasticsearch/repositories/gcs/GoogleCloudStorageBlobStoreRepositoryTests.java
@@ -24,16 +24,10 @@ import com.google.cloud.http.HttpTransportOptions;
 import com.google.cloud.storage.StorageOptions;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
-import org.apache.http.HttpStatus;
-import org.apache.lucene.util.ArrayUtil;
+import fixture.gcs.GoogleCloudStorageHttpHandler;
 import org.elasticsearch.cluster.metadata.RepositoryMetaData;
-import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.SuppressForbidden;
-import org.elasticsearch.common.UUIDs;
-import org.elasticsearch.common.bytes.BytesArray;
-import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.common.io.Streams;
-import org.elasticsearch.common.regex.Regex;
 import org.elasticsearch.common.settings.MockSecureSettings;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.ByteSizeUnit;
@@ -43,37 +37,20 @@ import org.elasticsearch.env.Environment;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.repositories.Repository;
 import org.elasticsearch.repositories.blobstore.ESMockAPIBasedRepositoryIntegTestCase;
-import org.elasticsearch.rest.RestStatus;
-import org.elasticsearch.rest.RestUtils;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.threeten.bp.Duration;
 
-import java.io.BufferedInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Locale;
 import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.elasticsearch.repositories.gcs.GoogleCloudStorageClientSettings.CREDENTIALS_FILE_SETTING;
 import static org.elasticsearch.repositories.gcs.GoogleCloudStorageClientSettings.ENDPOINT_SETTING;
 import static org.elasticsearch.repositories.gcs.GoogleCloudStorageClientSettings.TOKEN_URI_SETTING;
 import static org.elasticsearch.repositories.gcs.GoogleCloudStorageRepository.BUCKET;
 import static org.elasticsearch.repositories.gcs.GoogleCloudStorageRepository.CLIENT_NAME;
-import static org.hamcrest.Matchers.lessThanOrEqualTo;
 
 @SuppressForbidden(reason = "this test uses a HttpServer to emulate a Google Cloud Storage endpoint")
 public class GoogleCloudStorageBlobStoreRepositoryTests extends ESMockAPIBasedRepositoryIntegTestCase {
@@ -100,8 +77,8 @@ public class GoogleCloudStorageBlobStoreRepositoryTests extends ESMockAPIBasedRe
     @Override
     protected Map<String, HttpHandler> createHttpHandlers() {
         final Map<String, HttpHandler> handlers = new HashMap<>(2);
-        handlers.put("/", new InternalHttpHandler());
-        handlers.put("/token", new FakeOAuth2HttpHandler());
+        handlers.put("/", new GoogleCloudStorageHttpHandler("bucket"));
+        handlers.put("/token", new fixture.gcs.FakeOAuth2HttpHandler());
         return Collections.unmodifiableMap(handlers);
     }
 
@@ -206,187 +183,6 @@ public class GoogleCloudStorageBlobStoreRepositoryTests extends ESMockAPIBasedRe
                         };
                     }
                 });
-        }
-    }
-
-    /**
-     * Minimal HTTP handler that acts as a Google Cloud Storage compliant server
-     */
-    @SuppressForbidden(reason = "this test uses a HttpServer to emulate a Google Cloud Storage endpoint")
-    private static class InternalHttpHandler implements HttpHandler {
-
-        private final ConcurrentMap<String, BytesArray> blobs = new ConcurrentHashMap<>();
-
-        @Override
-        public void handle(final HttpExchange exchange) throws IOException {
-            final String request = exchange.getRequestMethod() + " " + exchange.getRequestURI().toString();
-            try {
-                if (Regex.simpleMatch("GET /storage/v1/b/bucket/o*", request)) {
-                    final Map<String, String> params = new HashMap<>();
-                    RestUtils.decodeQueryString(exchange.getRequestURI().getQuery(), 0, params);
-                    final String prefix = params.get("prefix");
-
-                    final List<Map.Entry<String, BytesArray>> listOfBlobs = blobs.entrySet().stream()
-                        .filter(blob -> prefix == null || blob.getKey().startsWith(prefix)).collect(Collectors.toList());
-
-                    final StringBuilder list = new StringBuilder();
-                    list.append("{\"kind\":\"storage#objects\",\"items\":[");
-                    for (Iterator<Map.Entry<String, BytesArray>> it = listOfBlobs.iterator(); it.hasNext(); ) {
-                        Map.Entry<String, BytesArray> blob = it.next();
-                        list.append("{\"kind\":\"storage#object\",");
-                        list.append("\"bucket\":\"bucket\",");
-                        list.append("\"name\":\"").append(blob.getKey()).append("\",");
-                        list.append("\"id\":\"").append(blob.getKey()).append("\",");
-                        list.append("\"size\":\"").append(blob.getValue().length()).append("\"");
-                        list.append('}');
-
-                        if (it.hasNext()) {
-                            list.append(',');
-                        }
-                    }
-                    list.append("]}");
-
-                    byte[] response = list.toString().getBytes(UTF_8);
-                    exchange.getResponseHeaders().add("Content-Type", "application/json; charset=utf-8");
-                    exchange.sendResponseHeaders(RestStatus.OK.getStatus(), response.length);
-                    exchange.getResponseBody().write(response);
-
-                } else if (Regex.simpleMatch("GET /storage/v1/b/bucket*", request)) {
-                    byte[] response = ("{\"kind\":\"storage#bucket\",\"name\":\"bucket\",\"id\":\"0\"}").getBytes(UTF_8);
-                    exchange.getResponseHeaders().add("Content-Type", "application/json; charset=utf-8");
-                    exchange.sendResponseHeaders(HttpStatus.SC_OK, response.length);
-                    exchange.getResponseBody().write(response);
-
-                } else if (Regex.simpleMatch("GET /download/storage/v1/b/bucket/o/*", request)) {
-                    BytesArray blob = blobs.get(exchange.getRequestURI().getPath().replace("/download/storage/v1/b/bucket/o/", ""));
-                    if (blob != null) {
-                        final String range = exchange.getRequestHeaders().getFirst("Range");
-                        Matcher matcher = Pattern.compile("bytes=([0-9]*)-([0-9]*)").matcher(range);
-                        assert matcher.find();
-
-                        byte[] response = Integer.parseInt(matcher.group(1)) == 0 ? blob.array() : new byte[0];
-                        exchange.getResponseHeaders().add("Content-Type", "application/octet-stream");
-                        exchange.sendResponseHeaders(RestStatus.OK.getStatus(), response.length);
-                        exchange.getResponseBody().write(response);
-                    } else {
-                        exchange.sendResponseHeaders(RestStatus.NOT_FOUND.getStatus(), -1);
-                    }
-
-                } else if (Regex.simpleMatch("DELETE /storage/v1/b/bucket/o/*", request)) {
-                    int deletions = 0;
-                    for (Iterator<Map.Entry<String, BytesArray>> iterator = blobs.entrySet().iterator(); iterator.hasNext(); ) {
-                        Map.Entry<String, BytesArray> blob = iterator.next();
-                        if (blob.getKey().equals(exchange.getRequestURI().toString())) {
-                            iterator.remove();
-                            deletions++;
-                        }
-                    }
-                    exchange.sendResponseHeaders((deletions > 0 ? RestStatus.OK : RestStatus.NO_CONTENT).getStatus(), -1);
-
-                } else if (Regex.simpleMatch("POST /batch/storage/v1", request)) {
-                    final String uri = "/storage/v1/b/bucket/o/";
-                    final StringBuilder batch = new StringBuilder();
-                    for (String line : Streams.readAllLines(new BufferedInputStream(exchange.getRequestBody()))) {
-                        if (line.length() == 0 || line.startsWith("--") || line.toLowerCase(Locale.ROOT).startsWith("content")) {
-                            batch.append(line).append('\n');
-                        } else if (line.startsWith("DELETE")) {
-                            final String name = line.substring(line.indexOf(uri) + uri.length(), line.lastIndexOf(" HTTP"));
-                            if (Strings.hasText(name)) {
-                                try {
-                                    final String blobName = URLDecoder.decode(name, UTF_8.name());
-                                    if (blobs.entrySet().removeIf(blob -> blob.getKey().equals(blobName))) {
-                                        batch.append("HTTP/1.1 204 NO_CONTENT").append('\n');
-                                        batch.append('\n');
-                                    }
-                                } catch (UnsupportedEncodingException e) {
-                                    batch.append("HTTP/1.1 404 NOT_FOUND").append('\n');
-                                    batch.append('\n');
-                                }
-                            }
-                        }
-                    }
-                    byte[] response = batch.toString().getBytes(UTF_8);
-                    exchange.getResponseHeaders().add("Content-Type", exchange.getRequestHeaders().getFirst("Content-Type"));
-                    exchange.sendResponseHeaders(RestStatus.OK.getStatus(), response.length);
-                    exchange.getResponseBody().write(response);
-
-                } else if (Regex.simpleMatch("POST /upload/storage/v1/b/bucket/*uploadType=multipart*", request)) {
-                    Optional<Tuple<String, BytesArray>> content = TestUtils.parseMultipartRequestBody(exchange.getRequestBody());
-                    if (content.isPresent()) {
-                        blobs.put(content.get().v1(), content.get().v2());
-
-                        byte[] response = ("{\"bucket\":\"bucket\",\"name\":\"" + content.get().v1() + "\"}").getBytes(UTF_8);
-                        exchange.getResponseHeaders().add("Content-Type", "application/json");
-                        exchange.sendResponseHeaders(RestStatus.OK.getStatus(), response.length);
-                        exchange.getResponseBody().write(response);
-                    } else {
-                        exchange.sendResponseHeaders(RestStatus.BAD_REQUEST.getStatus(), -1);
-                    }
-
-                } else if (Regex.simpleMatch("POST /upload/storage/v1/b/bucket/*uploadType=resumable*", request)) {
-                    final Map<String, String> params = new HashMap<>();
-                    RestUtils.decodeQueryString(exchange.getRequestURI().getQuery(), 0, params);
-                    final String blobName = params.get("name");
-                    blobs.put(blobName, BytesArray.EMPTY);
-
-                    byte[] response = Streams.readFully(exchange.getRequestBody()).utf8ToString().getBytes(UTF_8);
-                    exchange.getResponseHeaders().add("Content-Type", "application/json");
-                    exchange.getResponseHeaders().add("Location", httpServerUrl() + "/upload/storage/v1/b/bucket/o?"
-                        + "uploadType=resumable"
-                        + "&upload_id=" + UUIDs.randomBase64UUID()
-                        + "&test_blob_name=" + blobName); // not a Google Storage parameter, but it allows to pass the blob name
-                    exchange.sendResponseHeaders(RestStatus.OK.getStatus(), response.length);
-                    exchange.getResponseBody().write(response);
-
-                } else if (Regex.simpleMatch("PUT /upload/storage/v1/b/bucket/o?*uploadType=resumable*", request)) {
-                    final Map<String, String> params = new HashMap<>();
-                    RestUtils.decodeQueryString(exchange.getRequestURI().getQuery(), 0, params);
-
-                    final String blobName = params.get("test_blob_name");
-                    byte[] blob = blobs.get(blobName).array();
-                    assertNotNull(blob);
-
-                    final String range = exchange.getRequestHeaders().getFirst("Content-Range");
-                    final Integer limit = TestUtils.getContentRangeLimit(range);
-                    final int start = TestUtils.getContentRangeStart(range);
-                    final int end = TestUtils.getContentRangeEnd(range);
-
-                    final ByteArrayOutputStream out = new ByteArrayOutputStream();
-                    long bytesRead = Streams.copy(exchange.getRequestBody(), out);
-                    int length = Math.max(end + 1, limit != null ? limit : 0);
-                    assertThat((int) bytesRead, lessThanOrEqualTo(length));
-                    if (length > blob.length) {
-                        blob = ArrayUtil.growExact(blob, length);
-                    }
-                    System.arraycopy(out.toByteArray(), 0, blob, start, Math.toIntExact(bytesRead));
-                    blobs.put(blobName, new BytesArray(blob));
-
-                    if (limit == null) {
-                        exchange.getResponseHeaders().add("Range", String.format(Locale.ROOT, "bytes=%d/%d", start, end));
-                        exchange.getResponseHeaders().add("Content-Length", "0");
-                        exchange.sendResponseHeaders(308 /* Resume Incomplete */, -1);
-                    } else {
-                        assertThat(limit, lessThanOrEqualTo(blob.length));
-                        exchange.sendResponseHeaders(RestStatus.OK.getStatus(), -1);
-                    }
-                } else {
-                    exchange.sendResponseHeaders(RestStatus.INTERNAL_SERVER_ERROR.getStatus(), -1);
-                }
-            } finally {
-                exchange.close();
-            }
-        }
-    }
-
-    @SuppressForbidden(reason = "this test uses a HttpServer to emulate a fake OAuth2 authentication service")
-    private static class FakeOAuth2HttpHandler implements HttpHandler {
-        @Override
-        public void handle(final HttpExchange exchange) throws IOException {
-            byte[] response = ("{\"access_token\":\"foo\",\"token_type\":\"Bearer\",\"expires_in\":3600}").getBytes(UTF_8);
-            exchange.getResponseHeaders().add("Content-Type", "application/json");
-            exchange.sendResponseHeaders(HttpStatus.SC_OK, response.length);
-            exchange.getResponseBody().write(response);
-            exchange.close();
         }
     }
